+++ .github/workflows/all.yaml
@@ -22,7 +22,7 @@ jobs:
         env:
           NODE_AUTH_TOKEN: ${{ secrets.GHB_TOKEN }}
       - name: Lint
-        run: npm run lint:staged
+        run: npm run lint
       - name: Build Goerli
         run: npm run build:goerli
       - name: Build Mumbai
