+++ subgraph-mumbai.yaml
@@ -2,7 +2,6 @@ specVersion: 0.0.4
 schema:
   file: ./schema.graphql
 dataSources:
-
   # There is no "(Chronos) Traveler" on Mumbai
 
   - name: ERC7432-Immutable-Roles
